week06/explore_phila_data/requirements.txt
appnope==0.1.4
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.13.3
bleach==6.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
fiona==1.9.5
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.2
google-cloud-storage==3.1.0
google-crc32c==1.6.0.dev2
google-resumable-media==2.7.2
googleapis-common-protos==1.69.0
grpcio==1.70.0
grpcio-status==1.70.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.1.2
nbclient==0.10.1
nbconvert==7.16.6
nbformat==5.10.4
numpy==1.24.4
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
Pygments==2.19.1
pyproj==3.5.0
python-dateutil==2.9.0.post0
pyzmq==26.2.1
referencing==0.35.1
requests==2.31.0
rpds-py==0.20.1
rsa==4.9
shapely==2.0.3
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
tinycss2==1.2.1
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
zipp==3.20.2
